More robust reading of descriptors with trailing objects
The previous logic had hardcoded determinations to compute the trailing objects
for a particular class of descriptors.  As more descriptors gain more diverse
trailing objects, this gets harder to maintain.

This instead repeatedly estimates the full size and incrementally fetches more
data until the trailing object logic says we have everything.  It rests on a new
trailing object method `totalSizeOfPartialObject` that takes the size of a
(possibly partial) TrailingObjects type and returns either a smaller size
(indicating that the original data is sufficient) or a larger size (indicating
that the original data is insufficient).

Note that in the latter case, the returned size is necessarily an estimate;
because trailing objects can vary depending on the contents of earlier trailing
objects, we can't be entirely certain of the exact size until we have
everything.