Enable indexing for refs to implicit declarations.
Indexing was disabled for almost all references to implicit declarations, with an exception for constructors. This might be because `implicit` means the code is generated/synthesized at compile time and so the declaration doesn't "exist", in some sense. But the _reference_ very much exists and can be in non-synthesized code.

Enabling indexing for symbols more broadly will make the indexstore more useful and allow it to be used for more semantic analysis.

https://github.com/apple/swift/issues/67446