[CodeCompletion] Don’t report a call pattern as convertible if its result type doesn’t match
To compute the expected type of a call pattern (which is the return type of the function if that call pattern is being used), we called `getTypeForCompletion` for the entire call, in the same way that we do for the code completion token. However, this pattern does not generally work. For the code completion token it worked because the code completion expression doesn’t have an inherent type and it inherits the type solely from its context. Calls, however, have an inherent return type and that type gets assigned as the `typeForCompletion`.

To fix this, check if we recorded a fix that allowed a type mismatch between `ParentCall` and its surrounding context. If this is the case use the surrounding type as the `ExpectedCallType`. Finally, to make this work, we need to ensure that we don’t suppress the recording of such type mismatch fixes in code completion mode.