[Diagnostics] Add -no-warning-as-error to except a specific warning from being treated as an error
This commit adds a new compiler option -no-warning-as-error which allows users to specify exceptions from the -warnings-as-errors option.
The format of the option is -no-warning-as-error <diag_id>, where <diag_id> is one of the diagnostic IDs listed in include/swift/AST/Diagnostics*.def.
The option can be added multiple times.